ログインしてなくても見れるようにした

+++ app.rb
@@ -20,6 +20,9 @@ client = PG::connect(
 )
 
 get '/signup' do
+    if !session[:user].nil?
+        redirect '/'
+    end
     return erb :signup
 end
 
@@ -32,10 +35,13 @@ post '/signup' do
     client.exec_params("INSERT INTO users (name, email, password_digest, profile_img) VALUES ($1, $2, $3, $4)", [name, email, password_digest, profile_img])
     user = client.exec_params("SELECT * from users WHERE email = $1 AND password_digest = $2", [email, password_digest]).to_a.first
     session[:user] = user
-    return redirect '/home'
+    return redirect '/'
 end
 
 get '/user' do
+    if session[:user].nil?
+        redirect '/'
+    end
     @user = session[:user]
     # binding.pry
     return erb :user
@@ -55,12 +61,11 @@ post '/user_edit' do
     redirect '/user'
 end
 
-get '/' do
-    if session[:user].nil?
-        return erb :login
-    else
-        redirect '/home'
+get '/login' do
+    if !session[:user].nil?
+        redirect '/'
     end
+    return erb :login
 end
 
 post '/login' do
@@ -72,7 +77,7 @@ post '/login' do
         return erb :login
     else
         session[:user] = user
-        return redirect '/home'
+        return redirect '/'
     end
 end
 
@@ -81,12 +86,11 @@ delete '/logout' do
     return redirect '/'
 end
 
-get '/home' do
-    if session[:user].nil?
-        return redirect '/'
+get '/' do
+    if session[:user]
+        @user =  client.exec_params("SELECT * FROM users WHERE id = $1",[session[:user]['id']]).to_a.first
     end
     @today = Date.today.to_s
-    @user =  client.exec_params("SELECT * FROM users WHERE id = $1",[session[:user]['id']]).to_a.first
     @chats = client.exec_params("SELECT * FROM chats JOIN users ON chats.user_id = users.id ORDER BY chats.id DESC").to_a
     @schedules = client.exec_params("SELECT * FROM users JOIN schedules ON users.id = schedules.user_id ORDER BY date DESC").to_a
     # binding.pry
@@ -94,25 +98,34 @@ get '/home' do
 end
 
 post '/schedule_post' do
+    if session[:user].nil?
+        redirect '/login'
+    end
     user_id = session[:user]['id']
     subject = params[:subject]
     date = params[:date]
     session = params[:session]
     complement = params[:complement]
     client.exec_params("INSERT INTO schedules (user_id, subject, date, session, complement) VALUES ($1, $2, $3, $4, $5)",[user_id, subject, date, session, complement])
-    redirect '/home'
+    redirect '/'
 end
 
 delete '/schedule_delete/:id' do
+    if session[:user].nil?
+        redirect '/login'
+    end
     client.exec_params("DELETE FROM schedules WHERE id =$1",[params[:id]])
-    redirect '/home'
+    redirect '/'
 end
 
 post '/chat_post' do
+    if session[:user].nil?
+        redirect '/login'
+    end
     user_id = session[:user]["id"]
     content = params[:content]
     now = "now"
     binding.pry
     client.exec_params("INSERT INTO chats (user_id, content, created_at) VALUES ($1, $2, $3)",[user_id, content, 'now'])
-    redirect '/home'
+    redirect '/'
 end
